Rails

+++ db/migrate/20180508111730_create_items.rb
@@ -3,7 +3,7 @@ class CreateItems < ActiveRecord::Migration[5.2]
     create_table :items do |t|
       t.integer :user_id
       t.string :name
-      t.float :rental_price
+      t.decimal :rental_price
       
 
       t.timestamps
